update: remove old contract implementations (keeping everything that hasn't been ported over), aside for docs I think she ready.
