+++ rubi/rubi/__init__.py
@@ -1,5 +1,4 @@
-from .contracts_v2 import *
 from .network import *
-from .rubi import Rubicon
+from .contracts import *
 from .types import *
 from .client import Client
