--- rubi/rubi/book/__init__.py
@@ -1,1 +1,0 @@
-from .book import Book
