--- rubi/rubi/contracts/helper/__init__.py
@@ -1,2 +1,0 @@
-from .rolodex import networks 
-from .erc20 import ERC20, ERC20Signer
