--- rubi/rubi/contracts/helper/abis/ERC20.json
@@ -1,323 +1,0 @@
-[
-    {
-       "anonymous":false,
-       "inputs":[
-          {
-             "indexed":true,
-             "internalType":"address",
-             "name":"src",
-             "type":"address"
-          },
-          {
-             "indexed":true,
-             "internalType":"address",
-             "name":"guy",
-             "type":"address"
-          },
-          {
-             "indexed":false,
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"Approval",
-       "type":"event"
-    },
-    {
-       "anonymous":false,
-       "inputs":[
-          {
-             "indexed":true,
-             "internalType":"address",
-             "name":"dst",
-             "type":"address"
-          },
-          {
-             "indexed":false,
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"Deposit",
-       "type":"event"
-    },
-    {
-       "anonymous":false,
-       "inputs":[
-          {
-             "indexed":true,
-             "internalType":"address",
-             "name":"src",
-             "type":"address"
-          },
-          {
-             "indexed":true,
-             "internalType":"address",
-             "name":"dst",
-             "type":"address"
-          },
-          {
-             "indexed":false,
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"Transfer",
-       "type":"event"
-    },
-    {
-       "anonymous":false,
-       "inputs":[
-          {
-             "indexed":true,
-             "internalType":"address",
-             "name":"src",
-             "type":"address"
-          },
-          {
-             "indexed":false,
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"Withdrawal",
-       "type":"event"
-    },
-    {
-       "payable":true,
-       "stateMutability":"payable",
-       "type":"fallback"
-    },
-    {
-       "constant":true,
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          },
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          }
-       ],
-       "name":"allowance",
-       "outputs":[
-          {
-             "internalType":"uint256",
-             "name":"",
-             "type":"uint256"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "constant":false,
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"guy",
-             "type":"address"
-          },
-          {
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"approve",
-       "outputs":[
-          {
-             "internalType":"bool",
-             "name":"",
-             "type":"bool"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"nonpayable",
-       "type":"function"
-    },
-    {
-       "constant":true,
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          }
-       ],
-       "name":"balanceOf",
-       "outputs":[
-          {
-             "internalType":"uint256",
-             "name":"",
-             "type":"uint256"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "constant":true,
-       "inputs":[
-          
-       ],
-       "name":"decimals",
-       "outputs":[
-          {
-             "internalType":"uint8",
-             "name":"",
-             "type":"uint8"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "constant":false,
-       "inputs":[
-          
-       ],
-       "name":"deposit",
-       "outputs":[
-          
-       ],
-       "payable":true,
-       "stateMutability":"payable",
-       "type":"function"
-    },
-    {
-       "constant":true,
-       "inputs":[
-          
-       ],
-       "name":"name",
-       "outputs":[
-          {
-             "internalType":"string",
-             "name":"",
-             "type":"string"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "constant":true,
-       "inputs":[
-          
-       ],
-       "name":"symbol",
-       "outputs":[
-          {
-             "internalType":"string",
-             "name":"",
-             "type":"string"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "constant":true,
-       "inputs":[
-          
-       ],
-       "name":"totalSupply",
-       "outputs":[
-          {
-             "internalType":"uint256",
-             "name":"",
-             "type":"uint256"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "constant":false,
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"dst",
-             "type":"address"
-          },
-          {
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"transfer",
-       "outputs":[
-          {
-             "internalType":"bool",
-             "name":"",
-             "type":"bool"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"nonpayable",
-       "type":"function"
-    },
-    {
-       "constant":false,
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"src",
-             "type":"address"
-          },
-          {
-             "internalType":"address",
-             "name":"dst",
-             "type":"address"
-          },
-          {
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"transferFrom",
-       "outputs":[
-          {
-             "internalType":"bool",
-             "name":"",
-             "type":"bool"
-          }
-       ],
-       "payable":false,
-       "stateMutability":"nonpayable",
-       "type":"function"
-    },
-    {
-       "constant":false,
-       "inputs":[
-          {
-             "internalType":"uint256",
-             "name":"wad",
-             "type":"uint256"
-          }
-       ],
-       "name":"withdraw",
-       "outputs":[
-          
-       ],
-       "payable":false,
-       "stateMutability":"nonpayable",
-       "type":"function"
-    }
- ]
