--- rubi/rubi/contracts/helper/abis/MarketAidFactory.json
@@ -1,133 +1,0 @@
-[
-    {
-       "anonymous":false,
-       "inputs":[
-          {
-             "indexed":false,
-             "internalType":"address",
-             "name":"newInstance",
-             "type":"address"
-          }
-       ],
-       "name":"NotifyMarketAidSpawn",
-       "type":"event"
-    },
-    {
-       "inputs":[
-          
-       ],
-       "name":"admin",
-       "outputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          }
-       ],
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "inputs":[
-          
-       ],
-       "name":"createMarketAidInstance",
-       "outputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          }
-       ],
-       "stateMutability":"nonpayable",
-       "type":"function"
-    },
-    {
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"user",
-             "type":"address"
-          }
-       ],
-       "name":"getUserMarketAids",
-       "outputs":[
-          {
-             "internalType":"address[]",
-             "name":"",
-             "type":"address[]"
-          }
-       ],
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"market",
-             "type":"address"
-          }
-       ],
-       "name":"initialize",
-       "outputs":[
-          
-       ],
-       "stateMutability":"nonpayable",
-       "type":"function"
-    },
-    {
-       "inputs":[
-          
-       ],
-       "name":"initialized",
-       "outputs":[
-          {
-             "internalType":"bool",
-             "name":"",
-             "type":"bool"
-          }
-       ],
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "inputs":[
-          
-       ],
-       "name":"rubiconMarket",
-       "outputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          }
-       ],
-       "stateMutability":"view",
-       "type":"function"
-    },
-    {
-       "inputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          },
-          {
-             "internalType":"uint256",
-             "name":"",
-             "type":"uint256"
-          }
-       ],
-       "name":"userMarketAids",
-       "outputs":[
-          {
-             "internalType":"address",
-             "name":"",
-             "type":"address"
-          }
-       ],
-       "stateMutability":"view",
-       "type":"function"
-    }
- ]
