+++ rubi/rubi/contracts/router.py
@@ -1,256 +1,339 @@
-import hexbytes
-import logging as log
-from web3 import Web3 
-from eth_abi import decode
-from eth_abi.codec import ABICodec
-from web3._utils.events import get_event_data
+from typing import Optional, Tuple, List
 
-from rubi.contracts.helper import networks
+from eth_typing import ChecksumAddress
+from web3 import Web3
+from web3.contract import Contract
 
-class RubiconRouter: 
-    """this class represents a RubiconRouter.sol contract with read functionality.
-    
+from rubi.contracts.base_contract import BaseContract
+from rubi.network import Network
+
+
+class RubiconRouter(BaseContract):
+    """This class represents the RubiconRouter.sol contract and by default has read functionality.
+    If a wallet and key are passed in instantiation then this class can also be used to write to the contract instance.
+
     :param w3: Web3 instance
     :type w3: Web3
-    :param contract: an optional contract instance, if not provided, the contract will be instantiated using the address and abi from the networks.py file given the chain id of the w3 instance
-    :type contract: Web3 Object, optional
+    :param contract: Contract instance
+    :type contract: Contract
+    :param wallet: a wallet address of the signer (optional)
+    :type wallet: Optional[ChecksumAddress]
+    :param key: the private key of the signer (optional)
+    :type key: Optional[str]
     """
 
-    # init function
-    def __init__(self, w3, contract=None):
+    def __init__(
+        self,
+        w3: Web3,
+        contract: Contract,
+        wallet: Optional[ChecksumAddress] = None,
+        key: Optional[str] = None
+    ) -> None:
+        """constructor method"""
+        super().__init__(
+            w3=w3,
+            contract=contract,
+            wallet=wallet,
+            key=key
+        )
 
-        chain = w3.eth.chain_id
+    @classmethod
+    def from_network(
+        cls,
+        network: Network,
+        wallet: Optional[ChecksumAddress] = None,
+        key: Optional[str] = None
+    ) -> "RubiconRouter":
+        """Create a RubiconRouter instance based on a Network instance.
 
-        if contract:
-            self.contract = contract
-            self.address = self.contract.address
-        else:
-            network = networks[chain]()
-            self.contract = w3.eth.contract(address=network.router, abi=network.router_abi)
-            self.address = network.router
-
-        # set the class variables
-        self.w3 = w3
-        self.log_swap_abi = self.contract.events.LogSwap._get_event_abi()
-        self.codec: ABICodec = w3.codec
+        :param network: A Network instance.
+        :type network: Network
+        :param wallet: Optional wallet address to use for interacting with the contract.
+        :type wallet: Optional[ChecksumAddress]
+        :param key: Optional private key for the wallet.
+        :type key: Optional[str]
+        :return: A RubiconRouter instance based on the Network instance.
+        :rtype: RubiconRouter
+        """
+        return cls.from_address_and_abi(
+            w3=network.w3,
+            address=network.rubicon.router.address,
+            contract_abi=network.rubicon.router.abi,
+            wallet=wallet,
+            key=key
+        )
 
     ######################################################################
     # read calls
-    ######################################################################  
-    
-    # checkClaimAllUserBonusTokens(user (address), targetBathTokens (address[]), token (address))
-    def check_claim_all_user_bonus_tokens(self, user, targetBathTokens, token):
-        """check the claim the user has to bonus tokens 
+    ######################################################################
 
-        :param user: the user address of interest
-        :type user: str
-        :param targetBathTokens: the bath tokens of interest, this is an array of addresses
-        :type targetBathTokens: list
-        :param token: the bonus token of interest
-        :type token: str
-        :return: the amount of bonus tokens the user has available to claim
+    # getMakerBalance(baseToken (address), tokens (List[address]), maker (address))
+    # -> (uint256 balanceInBook, uint256 balance)
+    def get_maker_balance(
+        self,
+        base_token: ChecksumAddress,
+        tokens: List[ChecksumAddress],
+        maker: ChecksumAddress,
+    ) -> Tuple[int, int]:
+        """Iterates through all the base_token/tokens[i] offers of the maker and returns the balance of the base_token
+        in the book and the balance of the base token
+
+        :param base_token:
+        :type base_token: ChecksumAddress
+        :param tokens:
+        :type tokens: List[ChecksumAddress]
+        :param maker: ChecksumAddress
+        :type maker
+        :return: balance in book, total token balance
+        :rtype: Tuple[int, int]
+        """
+
+        return self.contract.functions.getMakerBalance(base_token, tokens, maker).call()
+
+    # getMakerBalanceInPair(asset (address), quote (address), maker (address)) -> (uint256 balance)
+    def get_maker_balance_in_pair(
+        self,
+        asset: ChecksumAddress,
+        quote: ChecksumAddress,
+        maker: ChecksumAddress,
+    ) -> int:
+        """Retrieves the balance of a specific maker for a given asset/quote pair.
+
+        :param asset: The address of the asset token.
+        :type asset: ChecksumAddress
+        :param quote: The address of the quote token.
+        :type quote: ChecksumAddress
+        :param maker: The address of the maker.
+        :type maker: ChecksumAddress
+        :return: The balance of the maker in the specified asset/quote pair.
         :rtype: int
         """
 
-        try: 
-            claimable = self.contract.functions.checkClaimAllUserBonusTokens(user, targetBathTokens, token).call()
-        except Exception as e:
-            log.error(e, exc_info=True)
-            return None
-        
-        return claimable
+        return self.contract.functions.getMakerBalanceInPair(asset, quote, maker).call()
 
-    # getBestOffer(asset (address), quote(address))
-    def get_best_offer(self, asset, quote):
-        """get the best offer for a given asset and quote pair
-        
-        :param asset: the asset address of interest
-        :type asset: str
-        :param quote: the quote address of interest
-        :type quote: str
-        :return: the best offer id for the given asset and quote pair
+    # getBookFromPair(asset (address), quote (address)) -> (uint256[3][] asks, uint256[3][] bids)
+    def get_book_from_pair(
+        self,
+        asset: ChecksumAddress,
+        quote: ChecksumAddress,
+    ) -> Tuple[List[List[int]], List[List[int]]]:
+        """Retrieves the order book for a specific asset/quote pair.
+
+        :param asset: The address of the asset token.
+        :type asset: ChecksumAddress
+        :param quote: The address of the quote token.
+        :type quote: ChecksumAddress
+        :return: A tuple containing two lists: asks and bids. Each list contains a sublist of length 3, representing
+        the order book entries in the following format (pay_amt, buy_amt, id). The asks list represents the orders
+        selling the asset, while the bids list represents the orders buying the asset.
+        :rtype: Tuple[List[List[int]], List[List[int]]]
+        """
+
+        return self.contract.functions.getBookFromPair(asset, quote).call()
+
+    # getBookDepth(tokenIn (address), tokenOut (address)) -> (uint256 depth, uint256 bestOfferID)
+    def get_book_depth(self, token_in: ChecksumAddress, token_out: ChecksumAddress) -> Tuple[int, int]:
+        """Retrieves the depth of one side of the order book for a specific token pair along with the id of the best
+        offer.
+
+        :param token_in: The address of the quote.
+        :type token_in: ChecksumAddress
+        :param token_out: The address of the asset.
+        :type token_out: ChecksumAddress
+        :return: A tuple containing the depth of the order book and the ID of the best offer for token_out/token_in.
+        :rtype: Tuple[int, int]
+        """
+
+        return self.contract.functions.getBookDepth(token_in, token_out).call()
+
+    # getBestOfferAndInfo(asset (address), quote (address)) -> (uint256 id, uint256, address, uint256, address)
+    def get_best_offer_and_info(
+        self,
+        asset: ChecksumAddress,
+        quote: ChecksumAddress
+    ) -> Tuple[int, int, ChecksumAddress, int, ChecksumAddress]:
+        """Retrieves the information and id of the best offer for a specific asset/quote pair.
+
+        :param asset: The address of the asset token.
+        :type asset: ChecksumAddress
+        :param quote: The address of the quote token.
+        :type quote: ChecksumAddress
+        :return: A tuple containing the ID of the best offer, the pay_amt, the address of the pay_gem,
+                 the buy_amt, and the address of the buy_gem.
+        :rtype: Tuple[int, int, ChecksumAddress, int, ChecksumAddress]
+        """
+
+        return self.contract.functions.getBestOfferAndInfo(asset, quote).call()
+
+    # getExpectedSwapFill(pay_amt (uint256), buy_amt_min (uint256), route (address[])) -> (uint256 amount)
+    def get_expected_swap_fill(self, pay_amt: int, buy_amt_min: int, route: List[ChecksumAddress]) -> int:
+        """Estimates the expected amount including fees when swapping the specified payment amount using the specified
+        route. reverts with an exception if the swap cannot achieve the buy_amt_min
+
+        :param pay_amt: The payment amount.
+        :type pay_amt: int
+        :param buy_amt_min: The minimum buy amount.
+        :type buy_amt_min: int
+        :param route: The route of addresses representing the swap path.
+        :type route: List[ChecksumAddress]
+        :return: The estimated swap amount including fees.
         :rtype: int
         """
 
-        try: 
-            best_offer = self.contract.functions.getBestOfferAndInfo(asset, quote).call()
-        except ValueError:
-            log.warning('most likely a checksum error... retrying with checksummed addresses')
-            asset = self.w3.to_checksum_address(asset)
-            quote = self.w3.to_checksum_address(quote)
-            best_offer = self.contract.functions.getBestOfferAndInfo(asset, quote).call()
-        except Exception as e:
-            log.error(e, exc_info=True)
-            return None
-        
-        return best_offer
+        return self.contract.functions.getExpectedSwapFill(pay_amt, buy_amt_min, route).call()
 
-    # getBookFromPair(asset (address), quote(address), topNOrders (uint256))
-    def get_book_from_pair(self, asset, quote, topNOrders):
-        """get the book for a given asset and quote pair, to the depth of topNOrders
-        
-        :param asset: the asset address of interest
-        :type asset: str
-        :param quote: the quote address of interest
-        :type quote: str
-        :param topNOrders: the depth of the book to return
-        :type topNOrders: int
-        :return: the book for the given asset and quote pair, returned as a sorted list of [[asset_amount, quote_amount, offer_id], ...]]]
-        :rtype: list
+    # getExpectedMultiswapFill(pay_amts (uint256[]), buy_amt_mins (uint256[]), routes (address[][]))
+    # -> (uint256 amount)
+    def get_expected_multiswap_fill(
+        self,
+        pay_amts: List[int],
+        buy_amt_mins: List[int],
+        routes: List[List[ChecksumAddress]]
+    ) -> int:
+        """Estimates the expected amount including fees when swapping multiple specified payment amount using multiple
+        specified routes. reverts with an exception if the multiswap cannot achieve the buy_amt_mins along each route
+
+        :param pay_amts: The list of payment amounts for each swap.
+        :type pay_amts: List[int]
+        :param buy_amt_mins: The list of minimum buy amounts for each swap.
+        :type buy_amt_mins: List[int]
+        :param routes: The list of routes, where each route is a list of addresses representing the swap path.
+        :type routes: List[List[ChecksumAddress]]
+        :return: The estimated multi-swap amount.
+        :rtype: int
         """
 
-        try: 
-            book = self.contract.functions.getBookFromPair(asset, quote, topNOrders).call()
-        except ValueError:
-            log.warning('most likely a checksum error... retrying with checksummed addresses')
-            asset = self.w3.to_checksum_address(asset)
-            quote = self.w3.to_checksum_address(quote)
-            book = self.contract.functions.getBookFromPair(asset, quote, topNOrders).call()
-        except Exception as e:
-            return None
-        
-        return book
+        return self.contract.functions.getExpectedMultiswapFill(pay_amts, buy_amt_mins, routes).call()
 
-    ######################################################################
-    # events & helpers
-    ######################################################################
+    # checkClaimAllUserBonusTokens(address user, address[] targetBathTokens, address token)
+    # -> (uint256 earnedAcrossPools)
+    def check_claim_all_user_bonus_tokens(
+        self,
+        user: ChecksumAddress,
+        target_bath_tokens: List[ChecksumAddress],
+        token: ChecksumAddress
+    ) -> int:
+        """Checks the all bonus tokens that can be claimed by a user earned across all specified rubicon pools.
 
-    # TODO: today the event signature is hardcoded, but we should be able to get it from the contract
-    # the graph does something similar to this when you run codegen, it should be a simple string manipulation problem from the abis 
-    def get_log_swap_hash(self): 
-        return self.w3.keccak(text="LogSwap(uint256,address,uint256,address,bytes32,uint256,address)").hex()
+        :param user: The address of the user.
+        :type user: ChecksumAddress
+        :param target_bath_tokens: The list of target bath tokens to claim bonus from.
+        :type target_bath_tokens: List[ChecksumAddress]
+        :param token: The address of the token for which the bonus is claimed.
+        :type token: ChecksumAddress
+        :return: The total amount earned across all pools.
+        :rtype: int
+        """
 
-    # TODO: determine if this is the right assumtption to make about how the data is being received 
-    # this assumes that the function is being used directly in the context of being passed raw data from a websocket stream that has been loaded and converted to an AttributeDict
-    # TODO: i feel like this could be done much faster... tracing will tell us
-    def stream_log_swap(self, data): 
-
-        # convert the topics, transaction hash, and block hash to hex strings
-        data['params']['result']['topics'] = [hexbytes.HexBytes(topic) for topic in data['params']['result']['topics']]
-        data['params']['result']['transactionHash'] = hexbytes.HexBytes(data['params']['result']['transactionHash'])
-        data['params']['result']['blockHash'] = hexbytes.HexBytes(data['params']['result']['blockHash'])
-
-        # get the event data from the log
-        try:  
-            event = get_event_data(self.codec, self.log_swap_abi, data['params']['result'])
-
-            # TODO: parse out the pair from the bytes encoding that is passed in the event (bytes32)
-            # now pass the swap back in the form of a dictionary
-            # TODO: this is probably not the most performant, we will optimize later
-            swap = {
-                    'txn': event['transactionHash'].hex(),
-                    'event': event['event'],
-                    'inputAmount': event['args']['inputAmount'],
-                    'inputERC20': event['args']['inputERC20'],
-                    'hurdleBuyAmtMin': event['args']['hurdleBuyAmtMin'],
-                    'targetERC20': event['args']['targetERC20'],
-                    'pair': event['args']['pair'],
-                    'realizedFill': event['args']['realizedFill'],
-                    'recipient': event['args']['recipient'],
-            }
-            return swap
-        
-        except Exception as e:
-            log.error(e, exc_info=True)
-            return None
-    
-    def parse_log_swap(self, log):
-
-        # get the event data from the log
-        try:
-            event = get_event_data(self.codec, self.log_swap_abi, log)
-
-            # decode the pair 
-            # TODO: this be broke 
-            pair = decode(['bytes32'], event['args']['pair'])[0]
-
-            # now pass the swap back in the form of a dictionary
-            # TODO: this is probably not the most performant, we will optimize later
-            swap = {
-                    'txn': event['transactionHash'].hex(),
-                    'event': event['event'],
-                    'inputAmount': event['args']['inputAmount'],
-                    'inputERC20': event['args']['inputERC20'],
-                    'hurdleBuyAmtMin': event['args']['hurdleBuyAmtMin'],
-                    'targetERC20': event['args']['targetERC20'],
-                    'pair': pair,
-                    'realizedFill': event['args']['realizedFill'],
-                    'recipient': event['args']['recipient'],
-            }
-            return swap
-        
-        except Exception as e:
-            log.error(e, exc_info=True)
-            return None
-
-class RubiconRouterSigner(RubiconRouter):
-    """this class represents a RubiconRouter.sol contract with read and write functionality. it is the super class of the RubiconRouter class, and inherints the read functionality from it.
-
-    :param w3: Web3 instance
-    :type w3: Web3
-    :param wallet: the signers wallet address 
-    :type wallet: str
-    :param key: the signers private key
-    :type key: str
-    :param contract: an optional contract instance, if not provided, the contract will be instantiated using the address and abi from the networks.py file given the chain id of the w3 instance
-    :type contract: Web3 Object, optional
-    """
-
-    def __init__(self, w3, wallet, key, contract=None):
-        super().__init__(w3, contract)
-        self.chain = w3.eth.chain_id
-        self.wallet = wallet
-        self.key = key
+        return self.contract.functions.checkClaimAllUserBonusTokens(user, target_bath_tokens, token).call()
 
     ######################################################################
     # write calls
     ######################################################################
 
-    # swap(pay_amt (uint256), buy_amt_min (uint256), route (address[]), expectedMarketFeeBPS (uint256))
-    def swap(self, pay_amt, buy_amt_min, route, expected_market_fee_bps=1, nonce=None, gas=3000000, gas_price=None):
-        """this function swaps the pay_amt of the first token in the route array for at least the buy_amt_min of the last token in the route array
+    # multiswap(address[][] routes, uint256[] pay_amts, uint256[] buy_amts_min, address to) -> bool
+    def multiswap(
+        self,
+        routes: List[List[ChecksumAddress]],
+        pay_amts: List[int],
+        buy_amts_min: List[int],
+        to: ChecksumAddress,
+        nonce: Optional[int] = None,
+        gas: int = 350000,
+        max_fee_per_gas: Optional[int] = None,
+        max_priority_fee_per_gas: Optional[int] = None
+    ) -> str:
+        """Perform a multiple swaps for the specified payment amounts using the specified routes. Reverts with an
+        exception if any of the swaps cannot achieve the buy_amt_min along the specified route.
 
-        :param pay_amt: the amount of the first token in the route array to swap, as an integer amount of the token's smallest unit
-        :type pay_amt: int
-        :param buy_amt_min: the minimum amount of the last token in the route array to receive, as an integer amount of the token's smallest unit
-        :type buy_amt_min: int
-        :param route: array of addresses representing the path of tokens to swap through
-        :type route: list
-        :param expected_market_fee_bps: the market fee, defaults to 1
-        :type expected_market_fee_bps: int, optional
-        :param nonce: nonce of the transaction, defaults to calling the chain state to get the nonce
-        :type nonce: int, optional
-        :param gas: gas limit of the transaction, defaults to a value of 3000000
-        :type gas: int, optional
-        :param gas_price: gas price of the transaction, defaults to the gas price of the chain
-        :type gas_price: int, optional
-        :return: the transaction object of the swap transaction, returns None if the transaction fails
-        :rtype: dict, None
+        :param routes: The list of routes, where each route is a list of addresses representing the swap path.
+        :type routes: List[List[ChecksumAddress]]
+        :param pay_amts: The list of payment amounts for each swap.
+        :type pay_amts: List[int]
+        :param buy_amts_min: The list of minimum buy amounts for each swap.
+        :type buy_amts_min: List[int]
+        :param to: The address of the recipient.
+        :type to: ChecksumAddress
+        :param nonce: Nonce of the transaction. Defaults to calling the chain state to get the nonce.
+        :type nonce: Optional[int]
+        :param gas: Gas limit of the transaction. Defaults to a very high estimate made when writing the class.
+        :type gas: int
+        :param max_fee_per_gas: Max fee that can be paid for gas. Defaults to max_priority_fee (from chain)
+        + (2 * base fee per gas of latest block).
+        :type max_fee_per_gas: Optional[int]
+        :param max_priority_fee_per_gas: Max priority fee that can be paid for gas.
+        Defaults to calling the chain to estimate the max_priority_fee_per_gas.
+        :type max_priority_fee_per_gas: Optional[int]
+        :return: Transaction hash.
+        :rtype: str
         """
 
-        if nonce is None:
-            txn_nonce = self.w3.eth.get_transaction_count(self.wallet)
-        else:
-            txn_nonce = nonce
+        multiswap = self.contract.functions.multiswap(routes, pay_amts, buy_amts_min, to)
 
-        if gas_price is None:
-            gas_price = self.w3.eth.gas_price
+        return self._default_transaction_handler(
+            instantiated_contract_function=multiswap,
+            gas=gas,
+            nonce=nonce,
+            max_fee_per_gas=max_fee_per_gas,
+            max_priority_fee_per_gas=max_priority_fee_per_gas
+        )
 
-        txn = {'chainId': self.chain, 'gas' : gas, 'gasPrice': gas_price, 'nonce': txn_nonce}
+    # swap(uint256 pay_amt, uint256 buy_amt_min, address[] memory route, address to) -> uint256
+    def swap(
+        self,
+        pay_amt: int,
+        buy_amt_min: int,
+        route: List[ChecksumAddress],
+        to: ChecksumAddress,
+        nonce: Optional[int] = None,
+        gas: int = 350000,
+        max_fee_per_gas: Optional[int] = None,
+        max_priority_fee_per_gas: Optional[int] = None
+    ) -> str:
+        """Perform a swap operation with the specified payment amount using the specified route and paying out to the
+        recipient. Reverts if the swap does not result in the buy_min_amount.
 
-        try: 
-            swap = self.contract.functions.swap(pay_amt, buy_amt_min, route, expected_market_fee_bps).build_transaction(txn)
-            swap = self.w3.eth.account.sign_transaction(swap, self.key)
-            self.w3.eth.send_raw_transaction(swap.rawTransaction)
+        :param pay_amt: The payment amount.
+        :type pay_amt: int
+        :param buy_amt_min: The minimum buy amount.
+        :type buy_amt_min: int
+        :param route: The route, represented as a list of addresses representing the swap path.
+        :type route: List[ChecksumAddress]
+        :param to: The address of the recipient.
+        :type to: ChecksumAddress
+        :param nonce: Nonce of the transaction. Defaults to calling the chain state to get the nonce.
+        :type nonce: Optional[int]
+        :param gas: Gas limit of the transaction. Defaults to a very high estimate made when writing the class.
+        :type gas: int
+        :param max_fee_per_gas: Max fee that can be paid for gas. Defaults to max_priority_fee (from chain)
+        + (2 * base fee per gas of latest block).
+        :type max_fee_per_gas: Optional[int]
+        :param max_priority_fee_per_gas: Max priority fee that can be paid for gas. Defaults to calling the chain to
+        estimate the max_priority_fee_per_gas.
+        :type max_priority_fee_per_gas: Optional[int]
+        :return: Transaction hash.
+        :rtype: str
+        """
 
-            # if a user is not providing a nonce, wait for the transaction to either be confirmed or rejected before continuing
-            if nonce is None: 
-                if self.w3.eth.wait_for_transaction_receipt(swap.hash)['status'] == 0:
-                    log.error(f'swap transaction {swap.hash.hex()} failed')
-                    raise SystemExit()
+        swap = self.contract.functions.swap(pay_amt, buy_amt_min, route, to)
 
-        except Exception as e:
-            log.error(e, exc_info=True)
-            return None
-        
-        return swap
+        return self._default_transaction_handler(
+            instantiated_contract_function=swap,
+            gas=gas,
+            nonce=nonce,
+            max_fee_per_gas=max_fee_per_gas,
+            max_priority_fee_per_gas=max_priority_fee_per_gas
+        )
+
+    # TODO
+    #  sellAllAmountForETH
+    #  sellAllAmountWithETH
+    #  buyAllAmountWithETH
+    #  buyAllAmountForETH
+    #  swapWithETH
+    #  swapForETH
+    #  offerWithETH
+    #  offerForETH
+    #  cancelForETH
+    #  depositWithETH
+    #  withdrawForETH
