+++ rubi/rubi/contracts/types/__init__.py
@@ -1,0 +1,1 @@
+from .events import *
