--- rubi/rubi/contracts_v2/__init__.py
@@ -1,4 +1,0 @@
-from .erc20 import ERC20
-from .market import RubiconMarket
-from .router import RubiconRouter
-from .types import *
