--- rubi/rubi/contracts_v2/types/__init__.py
@@ -1,2 +1,0 @@
-from .base_contract import BaseContract
-from .events import *
